Enable/disable grtp commands
Added /grtp enable
and /grtp disable

Need to re-write cooldown timer so it works in conjunction with the other mods. Currently, it resets the timer when it reaches the cooldown time. It needs to store the count timer value when it creates a new teleport point, and then count - startTimer = elapsedTime